requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
comtypes==1.1.14
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
gTTS==2.1.1
gTTS-token==1.1.3
httplib2==0.17.0
idna==2.8
mutagen==1.45.1
numpy==1.23.2
oauth2client==4.1.3
opencv-python==4.6.0.66
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pypiwin32==223
pyttsx3==2.90
pywin32==304
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
uritemplate==3.0.1
urllib3==1.25.8
